// src/api/adr/model/saved-query.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ADRQuery } from './adrquery';

/**
 * 
 * @export
 * @interface SavedQuery
 */
export interface SavedQuery {
    /**
     * 
     * @type {number}
     * @memberof SavedQuery
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof SavedQuery
     */
    'name'?: string;
    /**
     * 
     * @type {ADRQuery}
     * @memberof SavedQuery
     */
    'query'?: ADRQuery;
}